% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_impact.R
\name{add_impact}
\alias{add_impact}
\title{Return details of issues, with input an GitHub issue URL}
\usage{
add_impact(url_issue, vars = c("created_at", "type", "title", "comments",
  "state", "creator"), url_old_impact = NULL)
}
\description{
Return details of issues, with input an GitHub issue URL
}
\examples{
\dontrun{
url_issue = "https://github.com/mtennekes/tmap/issues/196"
geocompr:::add_impact(url_issue)
u = "https://github.com/Robinlovelace/geocompr/raw/master/our-impact.csv"
geocompr:::add_impact(url_issue, url_old_impact = u)
}
}
