% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-citations.R
\name{dl_citations}
\alias{dl_citations}
\title{Download citations
# Dependes on a zotero API key (e.g. stored in Sys.getenv("ZOTERO")):
# dl_citations(f = "refs.bib", 216746, Sys.getenv("ZOTERO"), collection = "VJS7CTCC")}
\usage{
dl_citations(f, user, collection, key = NULL)
}
\description{
Download citations
# Dependes on a zotero API key (e.g. stored in Sys.getenv("ZOTERO")):
# dl_citations(f = "refs.bib", 216746, Sys.getenv("ZOTERO"), collection = "VJS7CTCC")
}
