% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-chapter-code.R
\name{generate_book_stats}
\alias{generate_book_stats}
\title{Generate a data frame of book statistics per chapter}
\usage{
generate_book_stats(dir = ".")
}
\description{
Generate a data frame of book statistics per chapter
}
