% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-chapter-code.R
\name{generate_chapter_code}
\alias{generate_chapter_code}
\title{Extracts R code from each chapter and dumps it in the code folder}
\usage{
generate_chapter_code(dir = ".", out_dir = "code/chapters/")
}
\description{
Extracts R code from each chapter and dumps it in the code folder
}
