% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocompkg-package.R
\docType{package}
\name{geocompkg-package}
\alias{geocompkg}
\alias{geocompkg-package}
\title{geocompkg: Geocomputation with R Metapackage}
\description{
Supplement to the book Geocomputation with R (\url{https://geocompx.com/r}). The packages in the Imports are required to build the first chapter of the book. The packages in Suggests are required for Part II and III.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/geocompx/geocompkg}
  \item \url{https://geocompx.github.io/geocompkg/}
  \item Report bugs at \url{https://github.com/geocompx/geocompr/issues}
}

}
\author{
\strong{Maintainer}: Jakub Nowosad \email{nowosad.jakub@gmail.com}

Authors:
\itemize{
  \item Robin Lovelace
  \item Jannes Muenchow
}

}
\keyword{internal}
