% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_print_quality.R
\name{save_print_quality}
\alias{save_print_quality}
\title{# example for tmap
library(tmap)
m = qtm(world) +
  tm_shape(urban_agglomerations) +
  tm_dots(size = "population_millions") +
  tm_facets(by = "year")
f = "/tmp/urban-animated-print.png"}
\usage{
save_print_quality(m = NULL, f, width = 2000, height = 2000)
}
\description{
# example for tmap
library(tmap)
m = qtm(world) +
  tm_shape(urban_agglomerations) +
  tm_dots(size = "population_millions") +
  tm_facets(by = "year")
f = "/tmp/urban-animated-print.png"
}
