% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_read_sp.R
\name{st_read_sp}
\alias{st_read_sp}
\title{Fast reading of spatial data into Spatial classes}
\usage{
st_read_sp(dsn, ...)
}
\arguments{
\item{dsn}{The path of the data source}

\item{...}{Arguments passed to \code{sf::st_read()}}
}
\description{
Wrapper function to read a file with \code{\link[sf:st_read]{sf::st_read()}} and the convert the result to a \code{Spatial} data class
}
\examples{
x = st_read_sp(system.file("shapes/world.gpkg", package = "spData"))
class(x)
}
